chore(api-graphql): reorganize utils
