+++ packages/api-graphql/src/internals/utils/clientProperties/generateEnumsProperty.ts
