+++ packages/api-graphql/src/internals/utils/runtimeTypeGuards/isConfigureEventWithResourceConfig.ts
